refactor: extract qgb cli test suite
